requirements.txt
streamlit
openai
python-dotenv
langchain
chromadb
PyPDF2
pdfplumber
